// src/code.d.ts
export declare enum PackageType {
    /**握手 */
    shakehands = 0,
    /**心跳 */
    heartbeat = 1,
    /**消息 */
    data = 2
}
/**Socket 状态 */
export declare enum SocketStatus {
    /**打开 */
    OPEN = 0,
    /**正在握手 */
    SHAKING_HANDS = 1,
    /**握手完毕 */
    HANDSHAKE = 2,
    /**连接 */
    CONNECTION = 3,
    /**关闭 */
    CLOSE = 4,
    /**重连 */
    RECONNECTION = 5
}
export interface ShakehandsPackageData {
    id: string;
    ack: SocketStatus;
}
export interface PackageData {
    path: string;
    request_id: number;
    status?: number;
    msg?: string;
    data?: {
        [key: string]: any;
    } | number;
    [key: string]: any;
}